tests/wpt/web-platform-tests/app-history/navigate/goTo-cross-document.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<iframe id="i" src="/common/blank.html"></iframe>
<script>
async_test(t => {
  // Wait for after the load event so that the navigation doesn't get converted
  // into a replace navigation.
  window.onload = () => t.step_timeout(() => {
    assert_equals(i.contentWindow.appHistory.entries().length, 1);
    let next_step = "goto";
    let key = i.contentWindow.appHistory.current.key;
    i.contentWindow.appHistory.navigate("?1");
    i.onload = t.step_func(() => {
      if (next_step == "goto") {
        assert_equals(i.contentWindow.appHistory.entries().length, 2);
        assert_equals(i.contentWindow.appHistory.current, i.contentWindow.appHistory.entries()[1]);
        i.contentWindow.appHistory.goTo(key);
        next_step = "forward";
      } else if (next_step == "forward") {
        assert_equals(i.contentWindow.appHistory.entries().length, 2);
        assert_equals(key, i.contentWindow.appHistory.current.key);
        assert_equals(i.contentWindow.appHistory.current, i.contentWindow.appHistory.entries()[0]);
        assert_true(i.contentWindow.appHistory.canGoForward);
        i.contentWindow.appHistory.forward();
        next_step = "back";
      } else if (next_step == "back") {
        assert_equals(i.contentWindow.appHistory.entries().length, 2);
        assert_equals(i.contentWindow.appHistory.current, i.contentWindow.appHistory.entries()[1]);
        assert_true(i.contentWindow.appHistory.canGoBack);
        i.contentWindow.appHistory.back();
        next_step = "finish";
      } else if (next_step == "finish") {
        assert_equals(i.contentWindow.appHistory.entries().length, 2);
        assert_equals(i.contentWindow.appHistory.current, i.contentWindow.appHistory.entries()[0]);
        t.done();
      }
    });
  }, 0);
}, "cross-document navigate.goTo(), back(), and forward()");
</script>
